remove viewer, add security, bug fixed

+++ application/controllers/Kategori.php
@@ -43,10 +43,7 @@ class Kategori extends CI_Controller
 
 	public function edit()
 	{
-		if ($this->input->server('REQUEST_METHOD') !== 'POST') {
-			redirect($this->service_name, 'refresh');
-		}
-
+		if ($this->input->server('REQUEST_METHOD') !== 'POST') redirect($this->service_name);
 		$id_kategori = $this->input->post('id_kategori');
 
 		$data = [
@@ -56,29 +53,23 @@ class Kategori extends CI_Controller
 
 		$this->base_model->update($this->service_name, $data, $id_kategori);
 		set_toasts("Data $this->service_name berhasil diedit.", 'success');
-		redirect($this->service_name, 'refresh');
+		redirect($this->service_name);
 	}
 
 	public function delete($id_kategori = null)
 	{
 		if (is_null($id_kategori)) show_404();
 
-		$is_exist_arsip = !is_null($this->base_model->get_one_data_by('arsip', 'id_kategori', $id_kategori));
 		$is_exist_sub_kategori = !is_null($this->base_model->get_one_data_by('sub_kategori', 'id_kategori', $id_kategori));
-
-		if ($is_exist_arsip) {
-			set_toasts("Kategori tidak dapat dihapus dikarenakan kategori telah terpakai oleh data arsip", 'danger');
-			redirect($this->service_name, 'refresh');
-		}
 
 		if ($is_exist_sub_kategori) {
 			set_toasts("Kategori tidak dapat dihapus dikarenakan kategori telah terpakai oleh data sub kategori", 'danger');
-			redirect($this->service_name, 'refresh');
+			redirect($this->service_name);
 		}
 
 		$this->base_model->delete($this->service_name, $id_kategori);
 		set_toasts("Data $this->service_name berhasil di Hapus", 'success');
-		redirect($this->service_name, 'refresh');
+		redirect($this->service_name);
 	}
 
 	public function sub()
@@ -131,8 +122,16 @@ class Kategori extends CI_Controller
 	public function sub_delete($id_sub_kategori = null)
 	{
 		if (is_null($id_sub_kategori)) show_404();
+
+		$is_exist_arsip = !is_null($this->base_model->get_one_data_by('arsip', 'id_sub_kategori', $id_sub_kategori));
+
+		if ($is_exist_arsip) {
+			set_toasts("Kategori tidak dapat dihapus dikarenakan kategori telah terpakai oleh data arsip", 'danger');
+			redirect('kategori/sub');
+		}
+
 		$this->base_model->delete('sub_kategori', $id_sub_kategori);
 		set_toasts("Data Sub Kategori berhasil di Hapus", 'success');
-		redirect('kategori/sub', 'refresh');
+		redirect('kategori/sub');
 	}
 }
